Add a tpp_call slot, and add initial version of typeTppCall
