+++ from_cpython/Include/object.h
@@ -461,6 +461,7 @@ struct _typeobject {
     bool _flags[3];
     void* _tpp_descr_get;
     void* _tpp_hasnext;
+    void* _tpp_call;
 };
 
 /* The *real* layout of a type object when allocated on the heap */
